minor changes to branch.h and cut.h, apparently made long ago. Mostly changes from int to double. Added a couple members to cut.h.

+++ branch.h
@@ -18,16 +18,16 @@ public:
   /// Standard constructor
   branch(TString);
   branch(TString,TString);
-  branch(TString,TString,int,int,int);
-  branch(TString,TString,int,int,int,TString,TString);
+  branch(TString,TString,int,double,double);
+  branch(TString,TString,int,double,double,TString,TString);
   TString self;//what the branch is called in the tree
   TString name;//nickname--usually what you want to appear on a plot
   TString xlabel;
   TString ylabel;
   int nBins;
-  int loBin;
-  int hiBin;
-  void binning(int,int,int);
+  double loBin;
+  double hiBin;
+  void binning(int,double,double);
   vector<cut> c;//cuts to be applied to the branch
   void add_cut(TCut);
   void add_cut(TCut,TString);
@@ -44,14 +44,14 @@ branch::branch(TString temp1, TString temp2){
   self = temp1;
   name = temp2;
 }
-branch::branch(TString tself, TString tname, int nb, int lb, int hb){
+branch::branch(TString tself, TString tname, int nb, double lb, double hb){
   self = tself;
   name = tname;
   nBins = nb;
   loBin = lb;
   hiBin = hb;
 }
-branch::branch(TString tself, TString tname, int nb, int lb, int hb, TString x, TString y){
+branch::branch(TString tself, TString tname, int nb, double lb, double hb, TString x, TString y){
   self = tself;
   name = tname;
   nBins = nb;
@@ -60,7 +60,7 @@ branch::branch(TString tself, TString tname, int nb, int lb, int hb, TString x, 
   xlabel = x;
   ylabel = y;
 }
-void branch::binning(int b1, int b2, int b3){
+void branch::binning(int b1, double b2, double b3){
   nBins = b1;
   loBin = b2;
   hiBin = b3;
